Disable ActiveStorage in all environments
This allows file uploads to be attached to models. It is failing on
Heroku, because the default storage provider is the local disk, but
Heroku doesn't support that as it only has ephemeral storage.

To use this on Heroku, we would have to use an external file storage
provider like S3 on AWS, or AzureStorage, etc. We don't currently
need this functionality, so can disable it.