Change stub tenancy gateway to a builder class
The gateway class is now instead created on the fly. This allows us to create each individual class with different context (e.g. with certain tenancies.)

We could alternatively pass these in as stub-specific initialization args in some cases, but stub_const doesn't allow this as you're stubbing the class itself, not instances of the class.

We need this change for later ones.