Show all action diary, not just last ten
This commit gets actions from a different endpoint to the rest of the
tenancy record. Sinking this detail into the use case is desirable,
but I'm worried about attempting that with my current level of
understanding of the structure of this app, especially its testing.

This commit duplicates `stub_income_api_actions` across three spec
files, which isn't ideal. I've used `with_context` in the past for
this sort of stubbing, but I didn't want to make such large changes
in the associated PR.