Expand/collapse weeks w/GOV.UK details component
This commit replaces some custom javascript with the GOV.UK details
component, mirroring the use in the other scrollable view on these
pages. Some custom javascript does still remain to power this
functionality - it was tricky to unpick completely in a decent amount
of time.

Notably, the styling is partially broken - the icon is present, but
the text is not link-styled and does not highlight yellow on-click.
Suspect this is due to old versions of gems importing the GOV.UK
styles in this app.