Use correct key from API to fix 'send' visibility
This commit changes the lookup for finding out which letter is being
sent, as the app should behave differently if it's a 'letter before
action' - namely, it should not send via gov-notify.

It also corrects the corresponding response specs.

The original break (introduced in 534b43) highlights one of the
things that can go wrong with this feature being split across an app
boundary.