Send 'username' as filter to documents endpoint
This commit modifies the relevant controller, use case and gateway
to send the `current_user.name` through to the API, so that events
can be written to the action diary with a name alongside.

The use case seems anemic, but I did not want to spend time
addressing this in the interests of focus.

The documents controller is covered by specs that I think are less
than ideal, as they test internals / mock out internal classes.
This commit does not address the issues with these specs - it only
maintains them in a working state and cleans them up slightly.

This commit also corrects the use of `expect have_requested` in some
related specs, which is an assertion that always passes. The 'correct'
pattern - which at least fails if the implementation is missing - is
`expect(WebMmock).to have_requested...`.

Notably, there are other unrelated specs with similarly ineffectual
assertions elsewhere in the codebase that this commit does not
address.