Cannot guarantee action diary balance is correct
This is due to the fact we are calculating the balance from newest to 
oldest. Where the Action diary is recorded after the transaction has 
happened so the balance on the action diary has the correct balance, but 
we are not aware of the transaction as it happened before the action 
diary.