Rescue from ActionController::InvalidAuthenticityToken
This occurs when calling the create action in agreements controlle, potentically caused by the page caching. "If you use page caching, Rails will keep the static HTML for a page in the cache. After a while the authenticity token for this page is stale and posting a form will result in this error. In this case either don't use page caching, or disable the forgery protection for that controller action" [Source](https://appsignal.com/for/invalid_authenticity_token)